+++ src/constants.go
@@ -8,7 +8,7 @@ import (
 
 const (
 	// Current version
-	version = "0.10.5"
+	version = "0.10.6"
 
 	// Core
 	coordinatorDelayMax  time.Duration = 100 * time.Millisecond
@@ -46,4 +46,11 @@ const (
 	EvtSearchFin
 	EvtHeader
 	EvtClose
+)
+
+const (
+	exitOk        = 0
+	exitNoMatch   = 1
+	exitError     = 2
+	exitInterrupt = 130
 )