+++ src/curses/curses.go
@@ -3,7 +3,9 @@ package curses
 /*
 #include <ncurses.h>
 #include <locale.h>
-#cgo LDFLAGS: -lncurses
+#cgo darwin LDFLAGS: -lncurses
+#cgo linux,386 LDFLAGS: -lncurses
+#cgo linux,amd64 LDFLAGS: -l:libncurses.a -l:libtinfo.a -l:libgpm.a -ldl
 */
 import "C"
 
@@ -261,7 +263,7 @@ func Init(theme *ColorTheme, black bool, mouse bool) {
 	_screen = C.newterm(nil, C.stderr, C.stdin)
 	if _screen == nil {
 		fmt.Println("Invalid $TERM: " + os.Getenv("TERM"))
-		os.Exit(1)
+		os.Exit(2)
 	}
 	C.set_term(_screen)
 	if mouse {
@@ -275,7 +277,7 @@ func Init(theme *ColorTheme, black bool, mouse bool) {
 	go func() {
 		<-intChan
 		Close()
-		os.Exit(1)
+		os.Exit(2)
 	}()
 
 	if theme != nil {